Expose zkey to martian_atmosphere_api, which defines the altitude of the profile
The current default is zkey=2, which implies the altitude is the altitude above the Martian zero datum (Mars geoid or “areoid”/MOLA_0) (km). However, for some applications the altitude above local surface (zkey=3) is preferable. I have exposed them all here, in case we ever need them, but 1, 4 and 5 are not appropriate for the current fcm.